improve theory part
the beginning is very close to my original text
please leave this way of introducing the theory. Like this it is not connected
to valueing the different approaches and their importance.
Also the different parameters of the later calculation are introduced one by one and not all at the same time.